don't create d:file elements for directories when unzipping
It doesn't work properly with px:fileset-store.

Alternatively, we could fix px:fileset-store so that it
handles directories properly. That way, we would have a
way to represent empty directories in a d:fileset.

If we don't want to support handling empty directories
in d:filesets, please merge this PR.