Use DS annotations instead of XML everywhere
Also make the _dsannotations instructions explicit everywhere. This
makes it immediately clear which services a bundle exposes, which is
especially useful for the generated classes. It is also needed in order
to not include services in the same packages in other bundles.

_dsannotations is specified through a property "expose-services", so
that it can be reused later for the ds-to-spi-maven-plugin.

Also fix "split-package" warnings by adding Private-Package
instructions with a "split-package" directive. For the sake of clarity
make Private-Package explicit everywhere.

Also change pf:file-expand83 so that it doesn't affect non-existing
files.