fileset-utils: remove mock folder
because it is not used, and because there was already another mock
folder present (not identical).

Adding the mock folder (commit c95c3505efa6ba3c8da7e1c5c98b5ca072c965bb)
was probably a mistake because it contains fileset-utils.xpl which does
not make sense.

The other mock folder is also not used (the XProc part). If we would
like to use it we have to add a catalog.xml file and switch to Calabash
as the XProc engine.